Aggregator: Prepare for new handlers
