+++ pkg/aggregator/apiserver/plugin/query.go
@@ -1,4 +1,4 @@
-package apiserver
+package plugin
 
 import (
 	"encoding/json"
@@ -8,73 +8,21 @@ import (
 
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
 	data "github.com/grafana/grafana-plugin-sdk-go/experimental/apis/data/v0alpha1"
+	grafanasemconv "github.com/grafana/grafana/pkg/semconv"
 	semconv "go.opentelemetry.io/otel/semconv/v1.21.0"
 	"k8s.io/component-base/tracing"
 	"k8s.io/klog/v2"
 
 	aggregationv0alpha1 "github.com/grafana/grafana/pkg/aggregator/apis/aggregation/v0alpha1"
-	grafanasemconv "github.com/grafana/grafana/pkg/semconv"
+	"github.com/grafana/grafana/pkg/aggregator/apiserver/util"
 )
 
-type pluginHandler struct {
-	mux      *http.ServeMux
-	delegate http.Handler
-
-	client                PluginClient
-	pluginContextProvider PluginContextProvider
-
-	dataplaneService aggregationv0alpha1.DataPlaneService
-}
-
-func newPluginHandler(
-	client PluginClient,
-	dataplaneService aggregationv0alpha1.DataPlaneService,
-	pluginContextProvider PluginContextProvider,
-	delegate http.Handler,
-) *pluginHandler {
-	mux := http.NewServeMux()
-	h := &pluginHandler{
-		mux:                   mux,
-		delegate:              delegate,
-		client:                client,
-		pluginContextProvider: pluginContextProvider,
-		dataplaneService:      dataplaneService,
-	}
-
-	for _, service := range dataplaneService.Spec.Services {
-		switch service.Type {
-		case aggregationv0alpha1.QueryServiceType:
-			proxyPath := fmt.Sprintf("/apis/%s/%s/namespaces/{namespace}/connections/{uid}/query", dataplaneService.Spec.Group, dataplaneService.Spec.Version)
-			mux.Handle(proxyPath, h.QueryDataHandler())
-		case aggregationv0alpha1.StreamServiceType:
-			// TODO: implement in future PR
-		case aggregationv0alpha1.AdmissionControlServiceType:
-			// TODO: implement in future PR
-		case aggregationv0alpha1.RouteServiceType:
-			// TODO: implement in future PR
-		case aggregationv0alpha1.ConversionServiceType:
-			// TODO: implement in future PR
-		case aggregationv0alpha1.DataSourceProxyServiceType:
-			// TODO: implement in future PR
-		}
-	}
-
-	// fallback to the delegate
-	mux.Handle("/", delegate)
-
-	return h
-}
-
-func (h *pluginHandler) ServeHTTP(w http.ResponseWriter, req *http.Request) {
-	h.mux.ServeHTTP(w, req)
-}
-
-func (h *pluginHandler) QueryDataHandler() http.HandlerFunc {
+func (h *PluginHandler) QueryDataHandler() http.HandlerFunc {
 	return func(w http.ResponseWriter, req *http.Request) {
 		ctx := req.Context()
 		span := tracing.SpanFromContext(ctx)
 		span.AddEvent("QueryDataHandler")
-		responder := &responder{w: w}
+		responder := &util.Responder{ResponseWriter: w}
 		dqr := data.QueryDataRequest{}
 		if err := json.NewDecoder(req.Body).Decode(&dqr); err != nil {
 			responder.Error(w, req, err)
